added graphql mutation to check price of pickup (currently without business logic)
