+++ src/main/java/com/irdaislakhuafa/garbagepickupapi/controllers/gql/pickup/PickupMutation.java
@@ -3,6 +3,7 @@ package com.irdaislakhuafa.garbagepickupapi.controllers.gql.pickup;
 import com.irdaislakhuafa.garbagepickupapi.models.entities.Pickup;
 import com.irdaislakhuafa.garbagepickupapi.models.gql.request.pickup.PickupRequest;
 import com.irdaislakhuafa.garbagepickupapi.models.gql.request.pickup.PickupUpdateRequest;
+import com.irdaislakhuafa.garbagepickupapi.models.gql.response.PickupCheckPriceResponse;
 import com.irdaislakhuafa.garbagepickupapi.services.PickupService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.graphql.data.method.annotation.Argument;
@@ -30,5 +31,10 @@ public class PickupMutation {
         final var result = this.pickupService.update(pickup);
         return result;
     }
-    
+
+    @SchemaMapping
+    public PickupCheckPriceResponse checkPrice() {
+        final var result = this.pickupService.checkPrice(0, 0, 0);
+        return result;
+    }
 }
