--- src/main/java/com/irdaislakhuafa/garbagepickupapi/models/gql/JwtTokenResponse.java
@@ -1,13 +1,0 @@
-package com.irdaislakhuafa.garbagepickupapi.models.gql;
-
-import lombok.*;
-import lombok.experimental.SuperBuilder;
-
-@Data
-@SuperBuilder
-@NoArgsConstructor
-@AllArgsConstructor
-public class JwtTokenResponse {
-	private String message;
-	private String token;
-}
