+++ src/main/java/com/irdaislakhuafa/garbagepickupapi/models/gql/response/PickupCheckPriceResponse.java
@@ -1,0 +1,14 @@
+package com.irdaislakhuafa.garbagepickupapi.models.gql.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Data
+@SuperBuilder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PickupCheckPriceResponse {
+    private int price;
+}
