+++ src/main/java/com/irdaislakhuafa/garbagepickupapi/services/impl/AuthServiceImpl.java
@@ -5,8 +5,8 @@ import com.irdaislakhuafa.garbagepickupapi.exceptions.custom.DataNotFound;
 import com.irdaislakhuafa.garbagepickupapi.exceptions.custom.InvalidPassword;
 import com.irdaislakhuafa.garbagepickupapi.exceptions.custom.UserNotAvailable;
 import com.irdaislakhuafa.garbagepickupapi.models.entities.User;
-import com.irdaislakhuafa.garbagepickupapi.models.gql.JwtTokenResponse;
 import com.irdaislakhuafa.garbagepickupapi.models.gql.request.user.UserLoginRequest;
+import com.irdaislakhuafa.garbagepickupapi.models.gql.response.JwtTokenResponse;
 import com.irdaislakhuafa.garbagepickupapi.repository.UserRepository;
 import com.irdaislakhuafa.garbagepickupapi.services.AuthService;
 import com.irdaislakhuafa.garbagepickupapi.services.JwtService;
